Grappling hook lazer effect
Added: Lazer effect for the grappling hook
Added: Linked playable character to Eoin's build
Added: FX_Lazer_system
Added: NE_Lazer
Added: NE_Lazer_Impact
Added: FX_Lazer_Impact_System

Note to prog/GD:

The system is set up in a way where all you have to do is to tell the end point of the particle system to attach itself to wherever it is needed for it to work as a grappling hook.

There are two public parameters. LazerEnd and LazerSystem. LazerSystem is the start point, and "LazerEnd" is the end point. You can change the colour and width of the particle, but those parameters aren't public, which means you have to go into the particle system yourself.

I made it this way so that plugging everything in doesn't get too confusing.