Bridge puzzle part 3
This is the final implementation of the bridge puzzle. I've now added the functionality of the blue power cell holder being disabled until each other cell has been placed, as well as setting up the correct scales for the power cells. I was unable to find any appropriate VFX or sounds, as the VFX are physically not in my branch, and I didn't want to use a sound from an already existing thing and have it not fit. However, I will speak to Matej during polishing.