Rock puzzle and switch/pressure plate puzzle fixing/polishing
The switch/pressure plate puzzle now features rock textures on the plates.

The rocks puzzle is now updated to feature textures, and the compatibility with the new weapon types, but as of right now (18/11/2022 at 3:30PM) it doesn't register any collision with the fire bomb and suction grenade. However, I told Cameron that we can fix this during polishing as he said he's quite busy with other stuff today.