Tutorial Production Light Pass
Added: All the necessary lighting and baked for production.

There is a known issue with the lightmap density for BSP objects because they aren't supposed to be used for anything other than blockouts.

The resolution can't be increased, no matter what I have tried, so at some point, we are going to have to create proper models for EVERYTHING. It's unfortunate.